refactoring

+++ src/part03/task04/Main.java
@@ -6,6 +6,7 @@ class Main {
     public static void main(String[] args) {
         int n = 200;
         BigInteger facFac = calcFacFac(n);
+
         System.out.println(facFac);
     }
 
@@ -13,7 +14,7 @@ class Main {
         BigInteger facFac = BigInteger.valueOf(1);
 
         for (int i = 2; i <= n; i++) {
-            facFac = facFac.multiply(BigInteger.valueOf(i * i));
+            facFac = facFac.multiply(BigInteger.valueOf((long) i * i));
         }
         return facFac;
     }
